requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
icalendar>=5.0.0
